[FIX] fixed shift-reduce reduce-reduce collision - shift-reduce collision prioritize shift - reduce-reduce collision prioritize by rule order - ID is now 32bit

+++ ISA-32 assembler/lexer.hpp
@@ -22,7 +22,7 @@
 
 namespace lexer_generator {
 	namespace graph {
-		using ID = unsigned long long;
+		using ID = unsigned int;
 		class UIDSet {
 		private:
 			std::set<ID> used;
